refactor: refactor paths
