+++ src/features/authentication/confirm-user/index.ts
@@ -1,0 +1,1 @@
+export { default as Confirmation } from './ui/Confirmation.vue'
