+++ src/features/authentication/index.ts
@@ -1,6 +1,5 @@
 export { useResendOneTimePassword } from './confirm-user/model/use-resend-one-time-password'
 export { useSendOneTimePassword } from './confirm-user/model/use-send-one-time-password'
-export { default as Confirmation } from './confirm-user/ui/Confirmation.vue'
 export { default as authDe } from './i18n/authDe'
 export { default as authEn } from './i18n/authEn'
 export { default as translationKeys } from './i18n/keys'
